+++ praetorian_ssh_proxy/hanlers/connection_handler.py
@@ -1,9 +1,7 @@
-import os
 import threading
 
 import paramiko
 from paramiko import AutoAddPolicy
-from praetorian_api_client.errors import ApiException
 
 from praetorian_ssh_proxy.errors import SshException
 from praetorian_ssh_proxy.hanlers.interactive_handler import InteractiveHandler
@@ -20,7 +18,6 @@ class ConnectionHandler(threading.Thread):
         try:
             session = paramiko.Transport(self._application.user_client)
             session.add_server_key(self._application.keyfile)
-            paramiko.util.log_to_file(os.path.join(self._application.LOG_DIR, 'filename.log'))
         except Exception as e:
             raise SshException(client=self._application.user_client, message=f'Caught exception: {str(e)}')
         return session
@@ -37,23 +34,17 @@ class ConnectionHandler(threading.Thread):
 
     def _connect_to_remote(self, remote):
         try:
-            ssh_service = self._application.api_client.service.list(
-                remote_id=remote.id, service_type='ssh'
-            )[0]
-        except ApiException as e:
-            raise SshException(client=self._application.user_client, message=e.message)
-
-        try:
             self._application.ssh_client = paramiko.SSHClient()
             self._application.ssh_client.set_missing_host_key_policy(AutoAddPolicy())
             self._application.ssh_client.connect(
-                hostname=ssh_service.variables.get('host'),
-                port=ssh_service.variables.get('port'),
-                username=ssh_service.variables.get('user'),
-                password=ssh_service.variables.get('password')
+                hostname=remote.host,
+                port=remote.port,
+                username=remote.user,
+                password=remote.password
             )
         except Exception as e:
             raise SshException(client=self._application.user_client, message=str(e))
+        self._application.ssh_client_connected = True
 
     def run(self):
         session = self._create_ssh_session()
@@ -74,7 +65,7 @@ class ConnectionHandler(threading.Thread):
             ).serve_remote_menu()
 
             self._connect_to_remote(self._application.remote_checker.remote)
-            self._application.channel.send('Successfully connected to Praetorian SSH Proxy Server.\r\n\r\n')
+            self._application.channel.send(b'Successfully connected to Praetorian SSH Proxy Server.\r\n\r\n')
 
             InteractiveHandler.create_from_channel(
                 client=self._application.user_client,